chore: migrate ai endpoints to use vercel's ai sdk (#1119)
