--- packages/root-cms/core/ai-chat.ts
@@ -1,616 +1,0 @@
-/**
- * Vercel AI SDK-based chat for the `/cms/ai` page.
- *
- * The chat runs as a lightweight proxy between the CMS UI and the underlying
- * model provider. Conversation history is persisted to Firestore so that users
- * can resume past chats.
- */
-import crypto from 'node:crypto';
-import {promises as fs} from 'node:fs';
-import path from 'node:path';
-import {createAnthropic} from '@ai-sdk/anthropic';
-import {createGoogleGenerativeAI} from '@ai-sdk/google';
-import {createOpenAI} from '@ai-sdk/openai';
-import {createOpenAICompatible} from '@ai-sdk/openai-compatible';
-import {RootConfig} from '@blinkk/root';
-import {
-  convertToModelMessages,
-  generateText,
-  LanguageModel,
-  stepCountIs,
-  streamText,
-  ToolSet,
-  UIMessage,
-} from 'ai';
-import {Timestamp} from 'firebase-admin/firestore';
-import {createCmsTools, createReadOnlyCmsTools} from './ai-tools.js';
-import {RootCMSClient} from './client.js';
-
-/** Filename of the project-level instructions file loaded into the AI prompt. */
-export const ROOT_MD_FILENAME = 'ROOT.md';
-
-/**
- * Provider type for an AI model. Use `openai-compatible` for any OpenAI-style
- * endpoint (e.g. local Ollama, vLLM, OpenRouter).
- */
-export type AiProvider =
-  | 'openai'
-  | 'openai-compatible'
-  | 'anthropic'
-  | 'google';
-
-/**
- * Configuration for a single chat model.
- *
- * Inspired by Ollama's `Modelfile` and LiteLLM's model config: each entry maps
- * a CMS-facing id to a provider, model id and credentials.
- */
-export interface AiModelConfig {
-  /** Stable id used by the CMS UI and stored alongside chat history. */
-  id: string;
-  /** Optional human-readable label rendered in the model picker. */
-  label?: string;
-  /** Optional description shown under the label. */
-  description?: string;
-  /** AI provider/family to route requests to. */
-  provider: AiProvider;
-  /**
-   * Provider-specific model id (e.g. `gpt-4o`, `claude-opus-4-5`,
-   * `gemini-2.5-pro`). Defaults to `id` if omitted.
-   */
-  modelId?: string;
-  /** API key for the provider. */
-  apiKey?: string;
-  /** Override the provider's base URL (required for `openai-compatible`). */
-  baseURL?: string;
-  /** Custom headers to send with each request. */
-  headers?: Record<string, string>;
-  /** Capabilities advertised to the UI. */
-  capabilities?: {
-    /** Whether the model can call tools. Defaults to `true`. */
-    tools?: boolean;
-    /** Whether the model can stream reasoning/thinking. Defaults to `false`. */
-    reasoning?: boolean;
-    /** Whether the model accepts image attachments. Defaults to `false`. */
-    attachments?: boolean;
-  };
-}
-
-/**
- * Full AI config registered on the cms plugin.
- */
-export interface AiConfig {
-  /** Models exposed in the model picker. The first entry is the default. */
-  models: AiModelConfig[];
-  /** Id of the default model. Defaults to the first model in `models`. */
-  defaultModel?: string;
-  /**
-   * Optional system prompt prepended to every conversation. If a `ROOT.md`
-   * file exists at the project root, its contents are appended to this
-   * prompt automatically.
-   */
-  systemPrompt?: string;
-  /** Maximum tool-loop steps before stopping. Defaults to 10. */
-  maxSteps?: number;
-}
-
-interface ChatRecord {
-  id: string;
-  createdBy: string;
-  createdAt: Timestamp;
-  modifiedAt: Timestamp;
-  modelId?: string;
-  title?: string;
-  messages: UIMessage[];
-}
-
-/** Resolves an `AiModelConfig` to an AI SDK `LanguageModel` instance. */
-export function resolveLanguageModel(model: AiModelConfig): LanguageModel {
-  const modelId = model.modelId || model.id;
-  switch (model.provider) {
-    case 'openai': {
-      const provider = createOpenAI({
-        apiKey: model.apiKey,
-        baseURL: model.baseURL,
-        headers: model.headers,
-      });
-      return provider(modelId);
-    }
-    case 'openai-compatible': {
-      if (!model.baseURL) {
-        throw new Error(
-          `model "${model.id}" requires a baseURL for provider "openai-compatible"`
-        );
-      }
-      const provider = createOpenAICompatible({
-        name: model.id,
-        baseURL: model.baseURL,
-        apiKey: model.apiKey,
-        headers: model.headers,
-      });
-      return provider(modelId);
-    }
-    case 'anthropic': {
-      const provider = createAnthropic({
-        apiKey: model.apiKey,
-        baseURL: model.baseURL,
-        headers: model.headers,
-      });
-      return provider(modelId);
-    }
-    case 'google': {
-      const provider = createGoogleGenerativeAI({
-        apiKey: model.apiKey,
-        baseURL: model.baseURL,
-        headers: model.headers,
-      });
-      return provider(modelId);
-    }
-    default: {
-      throw new Error(`unknown ai provider: ${(model as any).provider}`);
-    }
-  }
-}
-
-/** Strips secrets from `AiConfig` before sending to the browser. */
-export function serializeAiConfig(config: AiConfig) {
-  return {
-    defaultModel: config.defaultModel || config.models[0]?.id,
-    models: config.models.map((m) => ({
-      id: m.id,
-      label: m.label || m.id,
-      description: m.description,
-      provider: m.provider,
-      capabilities: {
-        tools: m.capabilities?.tools !== false,
-        reasoning: m.capabilities?.reasoning ?? false,
-        attachments: m.capabilities?.attachments ?? false,
-      },
-    })),
-  };
-}
-
-/** Returns the AI config registered on the CMS plugin, or `null`. */
-export function getAiConfig(rootConfig: RootConfig): AiConfig | null {
-  const cmsPlugin = rootConfig.plugins?.find((p) => p.name === 'root-cms') as
-    | {getConfig: () => {ai?: AiConfig}}
-    | undefined;
-  const ai = cmsPlugin?.getConfig().ai;
-  if (!ai || !Array.isArray(ai.models) || ai.models.length === 0) {
-    return null;
-  }
-  return ai;
-}
-
-/** Returns the model config matching `modelId`, or the default model. */
-export function findModel(
-  config: AiConfig,
-  modelId?: string
-): AiModelConfig | null {
-  if (modelId) {
-    const match = config.models.find((m) => m.id === modelId);
-    if (match) {
-      return match;
-    }
-  }
-  const defaultId = config.defaultModel || config.models[0]?.id;
-  return config.models.find((m) => m.id === defaultId) || null;
-}
-
-/**
- * Persists chat sessions to Firestore so that users can resume past
- * conversations from the chat history sidebar.
- */
-export class ChatStore {
-  cmsClient: RootCMSClient;
-  user: string;
-
-  constructor(cmsClient: RootCMSClient, user: string) {
-    this.cmsClient = cmsClient;
-    this.user = user;
-  }
-
-  collection() {
-    return this.cmsClient.db.collection(
-      `Projects/${this.cmsClient.projectId}/AiChats`
-    );
-  }
-
-  async createChat(options?: {
-    id?: string;
-    modelId?: string;
-  }): Promise<ChatRecord> {
-    const id = options?.id || crypto.randomUUID();
-    const now = Timestamp.now();
-    const record: ChatRecord = {
-      id,
-      createdBy: this.user,
-      createdAt: now,
-      modifiedAt: now,
-      modelId: options?.modelId,
-      messages: [],
-    };
-    await this.collection().doc(id).set(record);
-    return record;
-  }
-
-  async getChat(id: string): Promise<ChatRecord | null> {
-    const snap = await this.collection().doc(id).get();
-    if (!snap.exists) {
-      return null;
-    }
-    const data = snap.data() as ChatRecord;
-    if (data.createdBy !== this.user) {
-      return null;
-    }
-    return data;
-  }
-
-  async listChats(options?: {limit?: number}): Promise<ChatRecord[]> {
-    const limit = options?.limit ?? 50;
-    // Sort client-side to avoid requiring a Firestore composite index on
-    // (createdBy, modifiedAt).
-    const res = await this.collection()
-      .where('createdBy', '==', this.user)
-      .get();
-    const records = res.docs.map((d) => d.data() as ChatRecord);
-    records.sort((a, b) => {
-      const aMs = a.modifiedAt?.toMillis?.() ?? 0;
-      const bMs = b.modifiedAt?.toMillis?.() ?? 0;
-      return bMs - aMs;
-    });
-    return records.slice(0, limit);
-  }
-
-  async deleteChat(id: string): Promise<void> {
-    const chat = await this.getChat(id);
-    if (!chat) {
-      return;
-    }
-    await this.collection().doc(id).delete();
-  }
-
-  async updateMessages(
-    id: string,
-    messages: UIMessage[],
-    options?: {modelId?: string; title?: string}
-  ): Promise<void> {
-    const updates: Record<string, any> = {
-      messages,
-      modifiedAt: Timestamp.now(),
-    };
-    if (options?.modelId) {
-      updates.modelId = options.modelId;
-    }
-    if (options?.title) {
-      updates.title = options.title;
-    }
-    await this.collection().doc(id).update(updates);
-  }
-}
-
-/**
- * Reads the project's `ROOT.md` file, if present. Mirrors the convention used
- * by tools like `AGENTS.md` and `CLAUDE.md`: developers can drop a markdown
- * file at the project root to give Root AI extra context about site-specific
- * patterns or conventions.
- *
- * Returns the file contents (trimmed) or `null` if the file is missing or
- * unreadable.
- */
-export async function readRootMd(rootDir: string): Promise<string | null> {
-  const filePath = path.join(rootDir, ROOT_MD_FILENAME);
-  try {
-    const contents = await fs.readFile(filePath, 'utf8');
-    const trimmed = contents.trim();
-    return trimmed || null;
-  } catch (err: any) {
-    if (err && err.code === 'ENOENT') {
-      return null;
-    }
-    console.error(`failed to read ${ROOT_MD_FILENAME}:`, err);
-    return null;
-  }
-}
-
-/**
- * Builds the full system prompt by combining the configured/default base
- * prompt with the contents of `ROOT.md`, when present. The project-level
- * file is appended under a clearly delimited section so the model can
- * distinguish it from the framework-provided instructions.
- */
-export function buildSystemPrompt(
-  basePrompt: string,
-  rootMd: string | null
-): string {
-  if (!rootMd) {
-    return basePrompt;
-  }
-  return [
-    basePrompt,
-    '',
-    `The project includes a \`${ROOT_MD_FILENAME}\` file with site-specific`,
-    'instructions, conventions and context provided by the developer. Treat',
-    'these instructions as authoritative for this project and follow them',
-    'when responding or calling tools.',
-    '',
-    `<${ROOT_MD_FILENAME}>`,
-    rootMd,
-    `</${ROOT_MD_FILENAME}>`,
-  ].join('\n');
-}
-
-/** Derives a short title from the first user message (used as fallback). */
-export function deriveChatTitle(messages: UIMessage[]): string {
-  const first = messages.find((m) => m.role === 'user');
-  if (!first) {
-    return 'New chat';
-  }
-  const text = first.parts
-    .filter((p: any) => p.type === 'text')
-    .map((p: any) => p.text)
-    .join(' ')
-    .trim();
-  if (!text) {
-    return 'New chat';
-  }
-  return text.length > 60 ? `${text.slice(0, 57)}…` : text;
-}
-
-/**
- * Uses the AI model to generate a short summary title for the chat.
- * Falls back to `deriveChatTitle` if the generation fails.
- */
-export async function generateChatTitle(
-  model: LanguageModel,
-  messages: UIMessage[]
-): Promise<string> {
-  const fallback = deriveChatTitle(messages);
-  if (fallback === 'New chat') {
-    return fallback;
-  }
-  try {
-    const result = await generateText({
-      model,
-      system:
-        'Generate a short title (max 50 characters) summarizing the following conversation. ' +
-        'Return only the title text, no quotes or punctuation at the end.',
-      prompt: messages
-        .slice(0, 6)
-        .map((m) => {
-          const text = m.parts
-            .filter((p: any) => p.type === 'text')
-            .map((p: any) => p.text)
-            .join(' ');
-          return `${m.role}: ${text}`;
-        })
-        .join('\n'),
-      maxOutputTokens: 30,
-    });
-    const title = result.text
-      .trim()
-      .replace(/["']+$/g, '')
-      .replace(/^["']+/g, '');
-    if (title) {
-      return title.length > 80 ? `${title.slice(0, 77)}…` : title;
-    }
-  } catch (err) {
-    console.error('failed to generate chat title:', err);
-  }
-  return fallback;
-}
-
-export interface RunChatStreamOptions {
-  rootConfig: RootConfig;
-  cmsClient: RootCMSClient;
-  config: AiConfig;
-  model: AiModelConfig;
-  messages: UIMessage[];
-  chatId: string;
-  user: string;
-}
-
-/**
- * Builds a streaming response that proxies the model's UI message stream
- * directly to the client. Persists the final message list to Firestore once
- * the stream finishes.
- */
-export async function runChatStream(
-  options: RunChatStreamOptions
-): Promise<Response> {
-  const {rootConfig, model, config, messages, cmsClient, user, chatId} =
-    options;
-  const languageModel = resolveLanguageModel(model);
-  const tools: ToolSet =
-    model.capabilities?.tools === false ? {} : createCmsTools();
-
-  const basePrompt =
-    config.systemPrompt ||
-    [
-      'You are an assistant embedded in the Root CMS admin UI.',
-      'Help the user explore and edit content, answer questions about',
-      'the project, and use the provided tools to read and write CMS docs.',
-      'Be concise and use markdown for rich responses.',
-    ].join(' ');
-  const rootMd = await readRootMd(rootConfig.rootDir);
-  const systemPrompt = buildSystemPrompt(basePrompt, rootMd);
-
-  const modelMessages = await convertToModelMessages(messages, {tools});
-  const result = streamText({
-    model: languageModel,
-    system: systemPrompt,
-    messages: modelMessages,
-    tools,
-    stopWhen: stepCountIs(config.maxSteps ?? 10),
-  });
-
-  return result.toUIMessageStreamResponse({
-    sendReasoning: model.capabilities?.reasoning ?? false,
-    originalMessages: messages,
-    onFinish: async ({messages: finalMessages}) => {
-      const store = new ChatStore(cmsClient, user);
-      const title = await generateChatTitle(languageModel, finalMessages);
-      const updates = {
-        modelId: model.id,
-        title,
-      };
-      try {
-        // Firestore rejects `undefined` values, but the AI SDK frequently
-        // produces messages with `metadata: undefined`. Strip them before
-        // persisting.
-        await store.updateMessages(
-          chatId,
-          stripUndefined(finalMessages) as UIMessage[],
-          updates
-        );
-      } catch (err) {
-        console.error('failed to persist chat history:', err);
-      }
-    },
-  });
-}
-
-export interface RunEditObjectStreamOptions {
-  rootConfig: RootConfig;
-  config: AiConfig;
-  model: AiModelConfig;
-  messages: UIMessage[];
-  /** The JSON object the user is editing. Injected as context for the model. */
-  editData: unknown;
-}
-
-/**
- * Streaming variant used by the array-item "Edit with AI" diff-viewer flow.
- *
- * Differences from `runChatStream`:
- *
- * - Uses an edit-specific system prompt that injects the JSON the user is
- *   editing and the project's `root-cms.d.ts` types, and instructs the model
- *   to end every response with a fenced ```json block containing the
- *   complete proposed new JSON. The client extracts that block to populate
- *   the diff viewer.
- * - Tools are filtered to read-only (see `createReadOnlyCmsTools`). The user
- *   approves changes via the modal's "Save" button, so the model must not
- *   mutate Firestore directly.
- * - No Firestore persistence — edit sessions are ephemeral and don't show up
- *   in the user's chat history.
- */
-export async function runEditObjectStream(
-  options: RunEditObjectStreamOptions
-): Promise<Response> {
-  const {rootConfig, model, config, messages, editData} = options;
-  const languageModel = resolveLanguageModel(model);
-  const tools: ToolSet =
-    model.capabilities?.tools === false ? {} : createReadOnlyCmsTools();
-
-  const editPromptText = (await import('../shared/ai/prompts/edit.txt'))
-    .default;
-  // Strip the legacy `{"data": ..., "message": ...}` output spec from the
-  // bundled prompt — for the streaming flow we replace it with explicit
-  // instructions to emit a fenced ```json code block instead.
-  const promptParts: string[] = [
-    stripLegacyEditOutputSpec(editPromptText),
-    '',
-    'Output format:',
-    '- Begin with a brief 1-2 sentence message describing what you changed.',
-    '- Then emit a single fenced code block tagged ```json containing the',
-    '  COMPLETE proposed new JSON object (including any unmodified fields).',
-    '- The code block must be the LAST content in your response. The CMS UI',
-    '  parses it and shows the result in a diff viewer for the user to',
-    '  approve before saving.',
-    '',
-    'Tool policy:',
-    '- The available tools are READ-ONLY. Use them only when you need extra',
-    '  context (e.g. inspecting the schema or referencing other CMS docs).',
-    '- You MUST NOT attempt to call write tools (e.g. doc_set, doc_create,',
-    '  doc_updateField). The user approves and saves changes manually via',
-    '  the modal\'s Save button.',
-  ];
-
-  // Append the project's root-cms.d.ts type definitions if present so the
-  // model can validate its output against the actual schema.
-  try {
-    const rootCmsDefsPath = path.join(rootConfig.rootDir, 'root-cms.d.ts');
-    const rootCmsDefs = await fs.readFile(rootCmsDefsPath, 'utf8');
-    if (rootCmsDefs.trim()) {
-      promptParts.push(
-        '',
-        'Here is the `root-cms.d.ts` file for this project:',
-        '```',
-        rootCmsDefs,
-        '```'
-      );
-    }
-  } catch (err: any) {
-    if (err?.code !== 'ENOENT') {
-      console.error('failed to read root-cms.d.ts:', err);
-    }
-  }
-
-  // Inject the JSON being edited at the end of the system prompt.
-  promptParts.push(
-    '',
-    'The JSON you must edit is:',
-    '```json',
-    JSON.stringify(editData ?? {}, null, 2),
-    '```'
-  );
-
-  const basePrompt = promptParts.join('\n');
-  const rootMd = await readRootMd(rootConfig.rootDir);
-  const systemPrompt = buildSystemPrompt(basePrompt, rootMd);
-
-  const modelMessages = await convertToModelMessages(messages, {tools});
-  const result = streamText({
-    model: languageModel,
-    system: systemPrompt,
-    messages: modelMessages,
-    tools,
-    stopWhen: stepCountIs(config.maxSteps ?? 10),
-  });
-
-  return result.toUIMessageStreamResponse({
-    sendReasoning: model.capabilities?.reasoning ?? false,
-    originalMessages: messages,
-  });
-}
-
-/**
- * Removes the trailing legacy output specification block from `edit.txt`
- * (the `{"data": ..., "message": ...}` JSON envelope used by the Genkit-era
- * chat). The streaming flow uses fenced code blocks instead, which we
- * append after this function returns.
- */
-function stripLegacyEditOutputSpec(prompt: string): string {
-  const marker =
-    'Finally, when you provide your response, it MUST be structured';
-  const idx = prompt.indexOf(marker);
-  if (idx === -1) {
-    return prompt;
-  }
-  return prompt.slice(0, idx).trimEnd();
-}
-
-/**
- * Recursively removes `undefined` values from an object/array. Returns a new
- * structure; the input is not mutated. Used to clean payloads before sending
- * them to Firestore, which rejects `undefined` outright.
- */
-export function stripUndefined<T>(value: T): T {
-  if (Array.isArray(value)) {
-    return value
-      .filter((v) => v !== undefined)
-      .map((v) => stripUndefined(v)) as unknown as T;
-  }
-  if (value && typeof value === 'object') {
-    const out: Record<string, unknown> = {};
-    for (const [k, v] of Object.entries(value)) {
-      if (v === undefined) {
-        continue;
-      }
-      out[k] = stripUndefined(v);
-    }
-    return out as unknown as T;
-  }
-  return value;
-}
