+++ packages/root-cms/package.json
@@ -76,9 +76,6 @@
     "@ai-sdk/react": "3.0.177",
     "ai": "6.0.175",
     "zod": "4.4.3",
-    "@genkit-ai/ai": "1.26.0",
-    "@genkit-ai/core": "1.26.0",
-    "@genkit-ai/google-genai": "1.26.0",
     "@google-cloud/firestore": "7.11.3",
     "@hello-pangea/dnd": "18.0.1",
     "@types/cli-progress": "3.11.6",
@@ -93,7 +90,6 @@
     "dts-dom": "3.7.0",
     "@mantine/spotlight": "4.2.12",
     "fnv-plus": "1.3.1",
-    "genkit": "1.26.0",
     "jsonwebtoken": "9.0.2",
     "kleur": "4.1.5",
     "minisearch": "7.2.0",
